fix(matrix): the reference does not fuse the multiply-add
dotRow read d += a * b as a single fused multiply-add and rounded once. The
comment stated that as fact about what C++ emits; contraction is optional,
and the build the goldens come from does not do it — measured by mirroring
the whole hierarchical-softmax walk in numpy both ways and seeing which one
reproduced the reference. The product is rounded on its own now.

It was worth a lot: with the fused reading, 27 of 41 cases of lid.176.bin
were wrong, by up to 398 ulps. The quantized path is left alone, where a
single rounding is what the same reference build does — the two goldens pin
each of them.

None of this was visible, because the parity test compared to nine decimal
places with a tolerance of 1e-6, a hundred times looser than the 1e-5 floor
the float32 work exists for. Probabilities are now stored at full precision
and compared exactly, and mutating the arithmetic makes the suite red.

Exactly is not quite the whole truth, and the allowance says so: fastText
takes its exponentials in float, and the platform's expf is not always the
correctly rounded value, while Dart's rounded double exp is. Substituting
the platform's expf for every argument makes all 41 cases match to the bit,
which is how the residual was pinned down to three cases and eight ulps —
against 27 cases and 398 for a real mistake. Missing weights now fail rather
than skip when CI or LANG_ID_REQUIRE_MODELS says so, so the suite can no
longer report success on a clone that never ran its main test.

+++ lib/src/matrix.dart
@@ -56,11 +56,13 @@ class DenseMatrix implements Matrix {
   double dotRow(Float32List vector, int row) {
     final offset = row * columns;
     var sum = 0.0;
-    // The product is deliberately not rounded on its own: C++ contracts
-    // `d += a * b` into a single fused multiply-add, so there is exactly one
-    // rounding step, which is what storing a double product into float32 does.
+    // Two rounding steps, one for the product and one for the sum. Fusing
+    // them into a multiply-add would be the faster reading of `d += a * b`,
+    // and it is what the C++ is allowed to do, but the reference the goldens
+    // come from does not: contraction is optional, and the answers here were
+    // measured against a build that keeps both roundings.
     for (var i = 0; i < columns; i++) {
-      sum = float32(sum + vector[i] * _data[offset + i]);
+      sum = float32(sum + float32(vector[i] * _data[offset + i]));
     }
     return sum;
   }
@@ -217,7 +219,6 @@ class ProductQuantizer {
       final centroid = centroidOffset(m, codes[base + m]);
       final width = m == subquantizers - 1 ? lastSubDim : subDim;
       final offset = m * subDim;
-      // Single rounding, as in the fused multiply-add C++ emits here.
       for (var i = 0; i < width; i++) {
         target[offset + i] += scale * centroids[centroid + i];
       }
